+++ docs/api/phenotypes/factory/isth_major_bleed.md
@@ -1,0 +1,27 @@
+# Bleeding as defined by ISTH
+
+::: phenex.phenotypes.factory.isth_major_bleed.ISTHMajorBleedPhenotype
+
+# ISTH Bleed Components
+
+## Base class of bleed components
+
+::: phenex.phenotypes.factory.isth_major_bleed.ISTHBleedComponents
+
+
+# ISTH Major Bleed Sub Phenotypes
+
+## Critical Organ Bleed
+
+::: phenex.phenotypes.factory.isth_major_bleed.CriticalOrganBleedPhenotype
+
+## Symptomatic Bleed
+
+::: phenex.phenotypes.factory.isth_major_bleed.SymptomaticBleedPhenotype
+
+## Fatal Bleed
+
+::: phenex.phenotypes.factory.isth_major_bleed.FatalBleedPhenotype
+
+
+
