+++ phenex/phenotypes/factory/__init__.py
@@ -1,0 +1,1 @@
+from .isth_major_bleed import ISTHMajorBleedPhenotype
